% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lter_buoy_wtemp.R
\name{lter_buoy_wtemp}
\alias{lter_buoy_wtemp}
\title{LTER Buoy Water Temp}
\usage{
lter_buoy_wtemp(lake = "SP", name = "Anonymous Aardvark",
  email = "anonymous@1.edu", reason = "Academic+Research",
  start = "1950-01-01", end = "2020-01-01")
}
\description{
LTER Buoy Water Temp
}

