% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_ice_onoff.R
\name{lter_ice_onoff}
\alias{lter_ice_onoff}
\title{LTER lakes Northern Ice data}
\usage{
lter_ice_onoff()
}
\description{
Return the information on ice on/out for the LTER lakes.
}
\note{
Northern and southern ice on/off data are not described in exactly the same way, so some liberty was taken here. 
Southern data are reported only as "ice_on" and "ice_off". To match the northern data, I have renamed them 
"datelastopen" and "datefirstopen" respectively. I feel this is the closest analog.
}

