% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_north_chla.R
\name{lter_north_chla}
\alias{lter_north_chla}
\title{Get North LTER Chlorphyll Data}
\usage{
lter_north_chla(name = "Anonymous Aardvark", email = "anonymous@1.edu",
  reason = "Academic+Research", start = "1950-01-01", end = "2020-01-01")
}
\arguments{
\item{name}{Your name for LTER record keeping}

\item{email}{Your email address for LTER record keeping (you will not be unneccessarily contacted)}

\item{reason}{Your reason for usage of the data}

\item{start}{The earliest date to request data}

\item{end}{The latest date to request data}
}

