% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_POSIXct.R
\name{lter_POSIXct}
\alias{lter_POSIXct}
\title{Convert LTER date and numeric sampletime into POSIXct}
\usage{
lter_POSIXct(sampledate, sampletime, tz = "GMT")
}
\arguments{
\item{sampledate}{LTER sampledate, typically in YYYY-MM-DD format. Sometimes has additional 00:00:00 appended.}

\item{sampletime}{LTER sampletime, a numeric value where 100 is 1:00, 130 is 1:30, 30 is 0:30, and etc.}
}
\examples{

nice_r_datetime = lter_POSIXct('2014-01-01', 1230)

}

