% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_airport_meterology.R
\name{lter_airport_meterology}
\alias{lter_airport_meterology}
\title{Get high-resolution meterology data}
\usage{
lter_airport_meterology(name = "Anonymous Aardvark",
  email = "anonymous@1.edu", reason = "Academic+Research",
  start = "2000-01-01", end = "2001-01-01")
}
\arguments{
\item{name}{Your name for LTER record keeping}

\item{email}{Your email address for LTER record keeping (you will not be unneccessarily contacted)}

\item{reason}{Your reason for usage of the data}

\item{start}{The earliest date to request data}

\item{end}{The latest date to request data}
}
\description{
LTER Airport meterology station. Data are pulled the website here https://lter.limnology.wisc.edu/data/filter/10856
}
\note{
This query can be very slow. There is a large amount of high-res met-station data. For robustness, I recommend querying around a year at a time. Longer queries may time out.
}
\author{
Luke Winslow
}

