% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_secchi_basecrew.R
\name{lter_secchi_basecrew}
\alias{lter_secchi_basecrew}
\title{Get secchi and other standard base crew observed data}
\usage{
lter_secchi_basecrew()
}
\description{
Data from here https://lter.limnology.wisc.edu/dataset/north-temperate-lakes-lter-secchi-disk-depth-other-auxiliary-base-crew-sample-data-1981-curr
}

